% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_collapse.R
\name{act_collapse}
\alias{act_collapse}
\alias{act_collapse.eventlog}
\title{Collapse activity labels of a sub process into a single activity}
\usage{
act_collapse(eventlog, ...)

\method{act_collapse}{eventlog}(eventlog, ...)
}
\arguments{
\item{eventlog}{An \code{eventlog} object}

\item{...}{A series of named character vectors. The activity labels in each vector will be collapsed into one activity with the name of the vector.}
}
\description{
Collapse activity labels of a sub process into a single activity
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Collapse activity labels of a subprocess into a single activity
}}

\seealso{
Other Activity processing functions: \code{\link{act_recode}},
  \code{\link{act_unite}}
}
