% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_unite.R
\name{act_unite}
\alias{act_unite}
\alias{act_unite.log}
\alias{act_unite.grouped_log}
\title{Unite activity labels}
\usage{
act_unite(log, ...)

\method{act_unite}{log}(log, ...)

\method{act_unite}{grouped_log}(log, ...)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{...}{A series of named character vectors. The activity labels in each vector will be replaced with the name.}
}
\description{
Recode two or different more activity labels two a uniform activity label
}
\section{Methods (by class)}{
\itemize{
\item \code{act_unite(log)}: Unite activity labels in event log

\item \code{act_unite(grouped_log)}: Unite activity labels of event log

}}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}}, \code{\link{act_recode}}

Other Activity processing functions: 
\code{\link{act_collapse}()},
\code{\link{act_recode}()}
}
\concept{Activity processing functions}
