% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{activities}
\alias{activities}
\alias{activities.activitylog}
\alias{activities.grouped_log}
\title{Activities}
\usage{
activities(log, ...)

\method{activities}{activitylog}(log, ...)

\method{activities}{grouped_log}(log, ...)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{...}{Unused.}
}
\description{
Returns a \code{tibble}  containing a list of all activity types in the event log, with their absolute and relative frequency
}
\section{Methods (by class)}{
\itemize{
\item \code{activities(activitylog)}: Compute activity frequencies

\item \code{activities(grouped_log)}: Compute activity frequencies

}}
\seealso{
\code{\link{activity_id}},\code{\link{activity_instance_id}}, \code{\link{eventlog}}
}
