% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_id.R
\name{activity_id}
\alias{activity_id}
\title{Activity classifier}
\usage{
activity_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the activity classifier of an object of class \code{eventlog}.
}
\seealso{
\code{\link{eventlog}}, \code{\link{case_id}}, \code{\link{activity_instance_id}},
\code{\link{timestamp}}
}
