% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_instance_id.R
\name{activity_instance_id}
\alias{activity_instance_id}
\alias{activity_instance_id.eventlog}
\alias{activity_instance_id.eventlog_mapping}
\alias{activity_instance_id.activitylog}
\alias{activity_instance_id.activitylog_mapping}
\title{Activity instance classifier}
\usage{
activity_instance_id(x)

\method{activity_instance_id}{eventlog}(x)

\method{activity_instance_id}{eventlog_mapping}(x)

\method{activity_instance_id}{activitylog}(x)

\method{activity_instance_id}{activitylog_mapping}(x)
}
\arguments{
\item{x}{An \code{eventlog} of \code{eventlog_mapping}}
}
\description{
Get the activity instance classifier of an object of class \code{eventlog}.
}
\section{Methods (by class)}{
\itemize{
\item \code{activity_instance_id(eventlog)}: Retrieve activity instance identifier from eventlog

\item \code{activity_instance_id(eventlog_mapping)}: Retrieve activity instance identifier from eventlog mapping

\item \code{activity_instance_id(activitylog)}: Retrieve activity instance identifier from activitylog

\item \code{activity_instance_id(activitylog_mapping)}: Retrieve activity instance identifier from activitylog mapping

}}
\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
