% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_labels.R
\name{activity_labels}
\alias{activity_labels}
\alias{activity_labels.log}
\title{Get vector of activity labels}
\usage{
activity_labels(log)

\method{activity_labels}{log}(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Retrieve a vector containing all unique activity labels
}
\section{Methods (by class)}{
\itemize{
\item \code{activity_labels(log)}: Retrieve activity labels

}}
