% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_end_activity.R
\name{add_end_activity}
\alias{add_end_activity}
\alias{add_start_activity}
\alias{add_end_activity.eventlog}
\alias{add_end_activity.activitylog}
\alias{add_end_activity.grouped_log}
\alias{add_start_activity.eventlog}
\alias{add_start_activity.activitylog}
\alias{add_start_activity.grouped_log}
\title{Add Artificial Start/End Activities}
\usage{
add_end_activity(log, label = "End")

add_start_activity(log, label = "Start")

\method{add_end_activity}{eventlog}(log, label = "End")

\method{add_end_activity}{activitylog}(log, label = "End")

\method{add_end_activity}{grouped_log}(log, label = "End")

\method{add_start_activity}{eventlog}(log, label = "Start")

\method{add_start_activity}{activitylog}(log, label = "Start")

\method{add_start_activity}{grouped_log}(log, label = "Start")
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{label}{\code{\link{character}}: Start (default \code{"Start"}) or end (default \code{"End"}) activity label. This must be an activity
label that is not already present in \code{log}.}
}
\description{
Adds an artificial start or end activity to each case with the specified \code{label}.
}
\section{Methods (by class)}{
\itemize{
\item \code{add_end_activity(eventlog)}: Adds end activity to an \code{\link{eventlog}}.

\item \code{add_end_activity(activitylog)}: Adds end activity to an \code{\link{activitylog}}.

\item \code{add_end_activity(grouped_log)}: Adds end activity to a \code{\link{grouped_log}}.

}}
\section{Functions}{
\itemize{
\item \code{add_start_activity(eventlog)}: Adds start activity to an \code{\link{eventlog}}.

\item \code{add_start_activity(activitylog)}: Adds start activity to an \code{\link{activitylog}}.

\item \code{add_start_activity(grouped_log)}: Adds start activity to a \code{\link{grouped_log}}.

}}
