% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\docType{import}
\name{arrange}
\alias{arrange}
\title{Arrange event log}
\arguments{
\item{.data}{Eventlog}

\item{...}{Variables to arrange on}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}}
}}

