% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_helpers.R
\name{as.grouped.data.frame}
\alias{as.grouped.data.frame}
\title{as.grouped.data.frame}
\usage{
as.grouped.data.frame(data, groups)
}
\arguments{
\item{data}{Data}

\item{groups}{Names of grouping variables as character vector (e.g. by using \code{dplyr::group_vars}}
}
\description{
as.grouped.data.frame
}
