% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpar.R
\docType{package}
\name{bpaR}
\alias{bpaR}
\alias{bpaR-package}
\title{bpaR - Business Process Analysis in R}
\description{
bpaR is a collection of functions and other packages which can be used for (business) process analysis in R.
}

