% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupar.R
\docType{package}
\name{bupaR}
\alias{bupaR}
\title{bupaR - Business Process Analysis in R}
\description{
Functionalities for process analysis in R. This packages implements an S3-class for event log objects, and related handler functions. Imports related packages for subsetting event data, computation of descriptive statistics, handling of Petri Net objects and visualization of process maps.
}
