% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_id.r
\name{case_id}
\alias{case_id}
\alias{case_id.eventlog}
\alias{case_id.eventlog_mapping}
\alias{case_id.activitylog}
\alias{case_id.activitylog_mapping}
\title{Case classifier}
\usage{
case_id(x)

\method{case_id}{eventlog}(x)

\method{case_id}{eventlog_mapping}(x)

\method{case_id}{activitylog}(x)

\method{case_id}{activitylog_mapping}(x)
}
\arguments{
\item{x}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}, or \code{\link{mapping}}.}
}
\description{
Get the case classifier of an object of class \code{eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{case_id(eventlog)}: Retrieve case identifier from eventlog

\item \code{case_id(eventlog_mapping)}: Retrieve case identifier from eventlog mapping

\item \code{case_id(activitylog)}: Retrieve case identifier from activitylog

\item \code{case_id(activitylog_mapping)}: Retrieve case identifier from activitylog mapping

}}
\seealso{
\code{\link{eventlog}}, \code{\link{activitylog}}, \code{\link{mapping}}

Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
