% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamps.R
\name{convert_timestamps}
\alias{convert_timestamps}
\title{Convert timestamp format}
\usage{
convert_timestamps(x, columns, format = ymd_hms)
}
\arguments{
\item{x}{Data.frame containing events or activities.}

\item{columns}{A character vector with one or more names of columns to convert}

\item{format}{The format of the timestamps in the original dataset (either ymd_hms, dmy_hms, ymd_hm, ymd, dmy, dmy, ...). To be provided without quotation marks!}
}
\value{
Data.frame with converted timestamps
}
\description{
Function converting the timestamps in the data frame to the appropriate format.
}
\seealso{
Other Eventlog construction helpers: \code{\link{assign_instance_id}}
}
\concept{Eventlog construction helpers}
