% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\docType{import}
\name{count}
\alias{count}
\alias{count.log}
\alias{count.grouped_log}
\title{Count log}
\usage{
\method{count}{log}(x, ...)

\method{count}{grouped_log}(x, ...)
}
\arguments{
\item{x}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Variables to count on}
}
\section{Methods (by class)}{
\itemize{
\item \code{count(log)}: Count log

\item \code{count(grouped_log)}: Count grouped log

}}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{count}}}
}}

