% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_resource_inconsistencies.R
\name{detect_resource_inconsistencies}
\alias{detect_resource_inconsistencies}
\title{Detect resource inconsistencies}
\usage{
detect_resource_inconsistencies(eventlog, filter_condition)
}
\arguments{
\item{eventlog}{Event log object}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\description{
Function to detect inconsistencies in resource information between related events.
}
