% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durations.R
\name{durations}
\alias{durations}
\title{Durations}
\usage{
durations(
  log,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  eventlog = deprecated()
)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{units}{\code{\link{character}} (default "auto"): The time unit in which the throughput times should be reported. Should be one of the following values:
"auto" (default), "secs", "mins", "hours", "days", "weeks". See also the \code{units} argument of \code{\link{difftime}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Computes the throughput times of each case.
Throughput time is defined as the interval between the start of the first event and the completion of the last event.
}
