% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_to_activitylog.R
\name{events_to_activitylog}
\alias{events_to_activitylog}
\title{Events to activities}
\usage{
events_to_activitylog(eventlog)
}
\arguments{
\item{eventlog}{The event log to be converted. An object of class
\code{eventlog}.}
}
\description{
Events to activities
}
