% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.eventlog.R
\docType{import}
\name{fill}
\alias{fill}
\alias{fill.eventlog}
\alias{fill.grouped_eventlog}
\title{Fill event log}
\usage{
fill.eventlog(.data, ...)

fill.grouped_eventlog(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{options for fill}
}
\section{Functions}{
\itemize{
\item \code{fill.eventlog}: Fill eventlog

\item \code{fill.grouped_eventlog}: Fill grouped eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyr}{\code{\link[tidyr]{fill}}}
}}

