% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{import}
\name{filter}
\alias{filter}
\title{Filter event log}
\arguments{
\item{.data}{Eventlog}

\item{...}{Conditions to filter}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}
}}

