% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_attributes.R
\name{filter_attributes}
\alias{filter_attributes}
\alias{filter_attributes.eventlog}
\alias{filter_attributes.grouped_eventlog}
\alias{filter_attributes.activitylog}
\alias{filter_attributes.grouped_activitylog}
\title{Generic filter function for eventlog}
\usage{
filter_attributes(eventlog, ...)

\method{filter_attributes}{eventlog}(eventlog, ...)

\method{filter_attributes}{grouped_eventlog}(eventlog, ...)

\method{filter_attributes}{activitylog}(eventlog, ...)

\method{filter_attributes}{grouped_activitylog}(eventlog, ...)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{...}{Filter conditions}
}
\description{
Generic filter function for eventlog

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated as of bupaR version 0.5.0. Please use \link{filter} instead.
}
\section{Methods (by class)}{
\itemize{
\item \code{filter_attributes(eventlog)}: Filter eventlog using attributes

\item \code{filter_attributes(grouped_eventlog)}: Filter grouped eventlog using attributes

\item \code{filter_attributes(activitylog)}: Filter eventlog using attributes

\item \code{filter_attributes(grouped_activitylog)}: Filter grouped eventlog using attributes

}}
\keyword{internal}
