% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R
\name{group_by.eventlog}
\alias{group_by.eventlog}
\title{Group event log}
\usage{
\method{group_by}{eventlog}(.data, ..., add = F)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Variables to group by}

\item{add}{Add grouping variables to existing ones}
}
\description{
Group an event log
}
