% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_activity_instance.R
\name{group_by_activity_instance}
\alias{group_by_activity_instance}
\title{Group event log on activity instance id}
\usage{
group_by_activity_instance(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Group an event log by activity instance identifier
}
