% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_ids.R
\name{group_by_ids}
\alias{group_by_ids}
\alias{group_by_ids.log}
\title{Group log on identifiers}
\usage{
group_by_ids(.log, ...)

\method{group_by_ids}{log}(.log, ...)
}
\arguments{
\item{.log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{...}{One or more of the following: activity_id, case_id, activity_instance_id, resource_id, lifecycle_id}
}
\value{
Grouped log
}
\description{
Group log on identifiers
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Group log on identifiers
}}

