% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R, R/group_by_resource.R
\name{group_by_resource}
\alias{group_by_resource}
\alias{group_by_resource.eventlog}
\title{Group event log on resource id}
\usage{
group_by_resource(log)

\method{group_by_resource}{eventlog}(log)

group_by_resource(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}

\item{eventlog}{Eventlog}
}
\description{
Group an event log by resource identifier

Group an event log by resource identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on resource identifier
}}

