% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R,
%   R/group_by_resource_activity.R
\name{group_by_resource_activity}
\alias{group_by_resource_activity}
\alias{group_by_resource_activity.eventlog}
\title{Group event log on resource and activity id}
\usage{
group_by_resource_activity(log)

\method{group_by_resource_activity}{eventlog}(log)

group_by_resource_activity(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}

\item{eventlog}{Eventlog}
}
\description{
Group an event log by resource and activity identifier

Group an event log by resource and activity identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group an event log by resource and activity identifier
}}

