% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_trace.R
\name{group_by_trace}
\alias{group_by_trace}
\title{Group a log on trace}
\usage{
group_by_trace(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Group a log by trace
}
