% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.log}
\alias{is.log}
\alias{is.eventlog}
\alias{is.activitylog}
\alias{is.grouped_log}
\alias{is.grouped_eventlog}
\alias{is.grouped_activitylog}
\title{Test if the Object is a Log}
\usage{
is.log(x)

is.eventlog(x)

is.activitylog(x)

is.grouped_log(x)

is.grouped_eventlog(x)

is.grouped_activitylog(x)
}
\arguments{
\item{x}{Any \code{R} object.}
}
\value{
\code{is.log} returns \code{TRUE} if the object inherits from the \code{\link{log}} class, otherwise \code{FALSE}.

\code{is.eventlog} returns \code{TRUE} if the object inherits from the \code{\link{eventlog}} class, otherwise \code{FALSE}.

\code{is.actvitylog} returns \code{TRUE} if the object inherits from the \code{\link{activitylog}} class, otherwise \code{FALSE}.

\code{is.grouped_log} returns \code{TRUE} if the object inherits from the \code{\link{grouped_log}} class, otherwise \code{FALSE}.

\code{is.grouped_eventlog} returns \code{TRUE} if the object inherits from the \code{\link{grouped_eventlog}} class, otherwise \code{FALSE}.

\code{is.grouped_activitylog} returns \code{TRUE} if the object inherits from the \code{\link{grouped_activitylog}} class, otherwise \code{FALSE}.
}
\description{
This function returns \code{TRUE} if \code{x} inherits from the specified class, and \code{FALSE} for all other objects.
}
\seealso{
\code{\link{log}},\code{\link{eventlog}},\code{\link{activitylog}},\code{\link{grouped_log}},\code{\link{grouped_eventlog}},\code{\link{grouped_activitylog}}
}
