% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_id.R
\name{lifecycle_id}
\alias{lifecycle_id}
\alias{lifecycle_id.eventlog}
\alias{lifecycle_id.eventlog_mapping}
\title{Life cycle classifier}
\usage{
lifecycle_id(x)

\method{lifecycle_id}{eventlog}(x)

\method{lifecycle_id}{eventlog_mapping}(x)
}
\arguments{
\item{x}{\code{\link{eventlog}}: Object of class \code{\link{eventlog}}, or \code{\link{mapping}}.}
}
\description{
Get the life_cycle_id of an object of class \code{eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve lifecycle identifier from eventlog

\item \code{eventlog_mapping}: Retrieve lifecycle identifier from eventlog mapping
}}

\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
