% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_ids.R
\name{lifecycle_ids}
\alias{lifecycle_ids}
\alias{lifecycle_ids.eventlog}
\alias{lifecycle_ids.eventlog_mapping}
\alias{lifecycle_ids.activitylog}
\alias{lifecycle_ids.activitylog_mapping}
\title{Life cycle classifiers}
\usage{
lifecycle_ids(x)

\method{lifecycle_ids}{eventlog}(x)

\method{lifecycle_ids}{eventlog_mapping}(x)

\method{lifecycle_ids}{activitylog}(x)

\method{lifecycle_ids}{activitylog_mapping}(x)
}
\arguments{
\item{x}{An \code{eventlog} of \code{eventlog_mapping}}
}
\description{
Get the life_cycle_id of an object of class \code{activitylog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve lifecycle identifier from eventlog

\item \code{eventlog_mapping}: Retrieve lifecycle identifier from eventlog mapping

\item \code{activitylog}: Retrieve lifecycle identifier from activitylog

\item \code{activitylog_mapping}: Retrieve lifecycle identifier from activitylog mapping
}}

\seealso{
Other Eventlog classifiers: \code{\link{activity_id}},
  \code{\link{activity_instance_id}},
  \code{\link{case_id}}, \code{\link{lifecycle_id}},
  \code{\link{mapping}}, \code{\link{resource_id}},
  \code{\link{timestamp}}
}
\concept{Eventlog classifiers}
