% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycles.R
\name{lifecycle_labels}
\alias{lifecycle_labels}
\alias{lifecycle_labels.eventlog}
\title{Get vector of lifecycle labels}
\usage{
lifecycle_labels(eventlog)

\method{lifecycle_labels}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Retrieve a vector containing all unique lifecycle labels
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve lifecycle labels from eventlog
}}

