% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\alias{mapping.eventlog}
\alias{mapping.activitylog}
\title{Mapping}
\usage{
mapping(log, eventlog = deprecated())

\method{mapping}{eventlog}(log, eventlog = deprecated())

\method{mapping}{activitylog}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Prints the mapping of an event log object.
}
\section{Methods (by class)}{
\itemize{
\item \code{mapping(eventlog)}: Retrieve identifier mapping from eventlog

\item \code{mapping(activitylog)}: Retrieve identifier mapping from activitylog

}}
\concept{Eventlog classifiers}
