% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.eventlog.R
\name{mutate.eventlog}
\alias{mutate.eventlog}
\title{Mutate event log}
\usage{
\method{mutate}{eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{New variables}
}
\description{
Mutate an event log
}
