% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_activities.R
\name{n_activities}
\alias{n_activities}
\alias{n_activities.log}
\alias{n_activities.grouped_log}
\title{n_activities}
\usage{
n_activities(log)

\method{n_activities}{log}(log)

\method{n_activities}{grouped_log}(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Returns the number of activities in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{n_activities(log)}: Count the number of activities in a log

\item \code{n_activities(grouped_log)}: Count the number of activities for a grouped  log

}}
\seealso{
Other Counters: 
\code{\link{n_activity_instances}()},
\code{\link{n_cases}()},
\code{\link{n_events}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Counters}
