% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_activity_instances.R
\name{n_activity_instances}
\alias{n_activity_instances}
\alias{n_activity_instances.eventlog}
\alias{n_activity_instances.grouped_eventlog}
\alias{n_activity_instances.activitylog}
\alias{n_activity_instances.grouped_activitylog}
\title{n_activity_instances}
\usage{
n_activity_instances(log, eventlog = deprecated())

\method{n_activity_instances}{eventlog}(log, eventlog = deprecated())

\method{n_activity_instances}{grouped_eventlog}(log, eventlog = deprecated())

\method{n_activity_instances}{activitylog}(log, eventlog = deprecated())

\method{n_activity_instances}{grouped_activitylog}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Returns the number of activity instances in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{n_activity_instances(eventlog)}: eventlog

\item \code{n_activity_instances(grouped_eventlog)}: grouped_eventlog

\item \code{n_activity_instances(activitylog)}: eventlog

\item \code{n_activity_instances(grouped_activitylog)}: grouped_activitylog

}}
\seealso{
Other Counters: 
\code{\link{n_activities}()},
\code{\link{n_cases}()},
\code{\link{n_events}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Counters}
