% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cases.R
\name{n_cases}
\alias{n_cases}
\alias{n_cases.log}
\alias{n_cases.grouped_log}
\title{n_cases}
\usage{
n_cases(log, eventlog = deprecated())

\method{n_cases}{log}(log, eventlog = deprecated())

\method{n_cases}{grouped_log}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Returns the number of cases in an event log.
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Count number of cases in a \code{\link{log}}.

\item \code{grouped_log}: Count number of cases in a \code{\link{grouped_log}}.
}}

\seealso{
Other Counters: 
\code{\link{n_activities}()},
\code{\link{n_activity_instances}()},
\code{\link{n_events}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Counters}
