% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_resources.R
\name{n_resources}
\alias{n_resources}
\alias{n_resources.log}
\alias{n_resources.grouped_log}
\title{n_resources}
\usage{
n_resources(log, eventlog = deprecated())

\method{n_resources}{log}(log, eventlog = deprecated())

\method{n_resources}{grouped_log}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Returns the number of resources in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{n_resources(log)}: Count number of resources in log

\item \code{n_resources(grouped_log)}: Count number of resources in grouped log

}}
\seealso{
Other Counters: 
\code{\link{n_activities}()},
\code{\link{n_activity_instances}()},
\code{\link{n_cases}()},
\code{\link{n_events}()},
\code{\link{n_traces}()}
}
\concept{Counters}
