% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_traces.R
\name{n_traces}
\alias{n_traces}
\alias{n_traces.log}
\alias{n_traces.grouped_log}
\title{n_traces}
\usage{
n_traces(log, eventlog = deprecated(), ...)

\method{n_traces}{log}(log, eventlog = deprecated(), ...)

\method{n_traces}{grouped_log}(log, eventlog = deprecated(), ...)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Returns the number of traces in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Count number of traces for eventlog

\item \code{grouped_log}: Count number of traces for grouped eventlog
}}

\seealso{
Other Eventlog count functions: 
\code{\link{n_activities}()},
\code{\link{n_activity_instances}()},
\code{\link{n_cases}()},
\code{\link{n_events}()},
\code{\link{n_resources}()}
}
\concept{Eventlog count functions}
