% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eventlog.R
\name{print.eventlog}
\alias{print.eventlog}
\title{Generic print function for eventlog}
\usage{
\method{print}{eventlog}(x, ...)
}
\arguments{
\item{x}{Eventlog object}

\item{...}{Additional Arguments}
}
\description{
Generic print function for eventlog
}
