% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mapping.R
\name{print.eventlog_mapping}
\alias{print.eventlog_mapping}
\title{Generic print function for mapping.}
\usage{
\method{print}{eventlog_mapping}(x, ...)
}
\arguments{
\item{x}{Mapping of \code{\link{eventlog}} or \code{\link{activitylog}}}

\item{...}{Additional Arguments}
}
\description{
Generic print function for mapping.
}
