% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_map.R
\name{re_map}
\alias{re_map}
\title{Re map}
\usage{
re_map(x, mapping)
}
\arguments{
\item{x}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{mapping}{An existing mapping created by the mapping function}
}
\description{
Construct an eventlog using an existing mapping.
}
