% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.log}
\alias{rename.grouped_log}
\title{Rename log}
\usage{
rename(.data, ...)

\method{rename}{log}(.data, ...)

\method{rename}{grouped_log}(.data, ...)
}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Variables to rename. Use "new_name" = "old_name" to rename selected variables.}
}
\description{
Rename log
}
\section{Methods (by class)}{
\itemize{
\item \code{rename(log)}: Rename log

\item \code{rename(grouped_log)}: Rename grouped log

}}
