% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_labels.R
\name{resource_labels}
\alias{resource_labels}
\alias{resource_labels.default}
\title{Get vector of resource labels}
\usage{
resource_labels(log, eventlog = deprecated())

\method{resource_labels}{default}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Retrieve a vector containing all unique resource labels
}
\section{Methods (by class)}{
\itemize{
\item \code{resource_labels(default)}: Retrieve resource labels from eventlog

}}
