% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n.eventlog.R
\name{sample_n.eventlog}
\alias{sample_n.eventlog}
\title{Sample function for eventlog}
\usage{
\method{sample_n}{eventlog}(tbl, size, replace = FALSE, weight, .env)
}
\arguments{
\item{tbl}{Eventlog}

\item{size}{Number of cases to sample}

\item{replace}{Sample with replacement or not}

\item{weight}{N/A}

\item{.env}{N/A}
}
\description{
Return a sample of n cases of an event log object
}
