% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n.R
\name{sample_n}
\alias{sample_n}
\alias{sample_n.eventlog}
\alias{sample_n.grouped_eventlog}
\alias{sample_n.activitylog}
\title{Sample function for eventlog}
\usage{
sample_n(tbl, size, replace = FALSE, weight = NULL, .env, ...)

\method{sample_n}{eventlog}(tbl, size, replace = FALSE, weight, .env, ...)

\method{sample_n}{grouped_eventlog}(tbl, size, replace = FALSE, weight, .env, ...)

\method{sample_n}{activitylog}(
  log,
  tbl = deprecated(),
  size,
  replace = FALSE,
  weight = NULL,
  .env = deprecated()
)
}
\arguments{
\item{tbl}{Deprecated; please use \code{log} instead.}

\item{size}{\code{\link{integer}}: Number of cases to sample}

\item{replace}{\code{\link{logical}} (default \code{FALSE}): Sample with replacement \code{TRUE} or without \code{FALSE}.}

\item{weight}{Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input.
Weights are automatically standardised to sum to 1.}

\item{.env}{Deprecated; please don't use.}

\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}
}
\description{

}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Sample n cases of eventlog

\item \code{grouped_eventlog}: Stratified sampling of a grouped eventlog: sample n cases within each group

\item \code{activitylog}: Sample n cases from an \code{\link{activitylog}}.
}}

