% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ids.R
\name{select_ids}
\alias{select_ids}
\alias{select_ids.log}
\title{Select identifiers from log}
\usage{
select_ids(.log, ...)

\method{select_ids}{log}(.log, ...)
}
\arguments{
\item{.log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{...}{One or more of the following: activity_id, case_id, activity_instance_id, resource_id, lifecycle_id}
}
\description{
Select identifiers from log
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Select identifiers from log
}}

