% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_activity_id.R
\name{set_activity_id}
\alias{set_activity_id}
\alias{set_activity_id.default}
\title{Set activity id of log}
\usage{
set_activity_id(log, activity_id, eventlog = deprecated())

\method{set_activity_id}{default}(log, activity_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{activity_id}{New activity id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set activity id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Set activity id
}}

