% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_case_id.R
\name{set_case_id}
\alias{set_case_id}
\alias{set_case_id.default}
\title{Set case id of log}
\usage{
set_case_id(log, case_id, eventlog = deprecated())

\method{set_case_id}{default}(log, case_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{case_id}{New case id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set case id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Set case id
}}

