% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_resource_id.R
\name{set_resource_id}
\alias{set_resource_id}
\alias{set_resource_id.default}
\title{Set resource id of log}
\usage{
set_resource_id(log, resource_id, eventlog = deprecated())

\method{set_resource_id}{default}(log, resource_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{resource_id}{New resource id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set resource id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Set resource id
}}

