% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.eventlog.R
\name{slice.eventlog}
\alias{slice.eventlog}
\title{Slice function for event log}
\usage{
\method{slice}{eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Additional Arguments}
}
\description{
Return a slice of an event log object
}
