% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_activities.R
\name{slice_activities}
\alias{slice_activities}
\title{Slice Activities}
\usage{
slice_activities(.data, ...)
}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Slice index}
}
\description{
Take a slice of activity instances from event log
}
