% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_events.R
\name{slice_events}
\alias{slice_events}
\title{Slice Events}
\usage{
slice_events(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Slice index}
}
\description{
Take a slice of events from event log
}
