% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_lifecycle.R
\name{standardize_lifecycle}
\alias{standardize_lifecycle}
\alias{standardize_lifecycle.eventlog}
\title{Standardize format of lifecycle types}
\usage{
standardize_lifecycle(eventlog)

\method{standardize_lifecycle}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{The event log to be converted. An object of class
\code{eventlog}.}
}
\description{
Standardize format of lifecycle types
}
\section{Methods (by class)}{
\itemize{
\item \code{standardize_lifecycle(eventlog)}: Standardize lifecycle types for eventlog

}}
