% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\docType{import}
\name{summarize}
\alias{summarize}
\alias{summarise}
\title{Summarize event log}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}}

\item{...}{Name-value pairs of summary functions}
}
\value{
The summarize function returns a tible, no event log.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{summarise}}}
}}

