% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.eventlog.R
\name{summary.eventlog}
\alias{summary.eventlog}
\alias{summary.grouped_eventlog}
\title{Generic summary function for eventlog class}
\usage{
\method{summary}{eventlog}(object, ...)

\method{summary}{grouped_eventlog}(object, ...)
}
\arguments{
\item{object}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}}

\item{...}{Ignored.}
}
\description{
Generic summary function for eventlog class
}
\section{Methods (by class)}{
\itemize{
\item \code{summary(grouped_eventlog)}: Summary of grouped event log

}}
