% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\alias{timestamp.eventlog}
\alias{timestamp.eventlog_mapping}
\alias{timestamp.activitylog}
\alias{timestamp.activitylog_mapping}
\title{Timestamp classifier}
\usage{
timestamp(x)

\method{timestamp}{eventlog}(x)

\method{timestamp}{eventlog_mapping}(x)

\method{timestamp}{activitylog}(x)

\method{timestamp}{activitylog_mapping}(x)
}
\arguments{
\item{x}{Object of class \code{\link{eventlog}}, or \code{\link{mapping}}.}
}
\description{
Get the  timestamp classifier of an object of class \code{eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{timestamp(eventlog)}: Retrieve timestamp identifier from eventlog

\item \code{timestamp(eventlog_mapping)}: Retrieve timestamp identifier from eventlog mapping

\item \code{timestamp(activitylog)}: Retrieve timestamp identifier from activitylog

\item \code{timestamp(activitylog_mapping)}: Retrieve timestamp identifier from activitylog mapping

}}
\seealso{
\code{\link{eventlog}}, \code{\link{mapping}}

Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
