% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{timestamps}
\alias{timestamps}
\alias{timestamps.eventlog}
\alias{timestamps.eventlog_mapping}
\alias{timestamps.activitylog}
\alias{timestamps.activitylog_mapping}
\title{Timestamp classifiers}
\usage{
timestamps(x)

\method{timestamps}{eventlog}(x)

\method{timestamps}{eventlog_mapping}(x)

\method{timestamps}{activitylog}(x)

\method{timestamps}{activitylog_mapping}(x)
}
\arguments{
\item{x}{An \code{activitylog} or \code{activitylog_mapping}}
}
\description{
Get the  timestamps classifier of an object of class \code{activitylog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve timestamp identifier from eventlog

\item \code{eventlog_mapping}: Retrieve timestamp identifier from eventlog mapping

\item \code{activitylog}: Retrieve timestamp identifier from activitylog

\item \code{activitylog_mapping}: Retrieve timestamp identifier from activitylog mapping
}}

\seealso{
\code{\link{activitylog}}, \code{\link{mapping}}
}
\concept{Activitylog classifiers}
