% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_activitylog.R
\name{to_activitylog}
\alias{to_activitylog}
\title{Convert eventlog object to activitylog object.}
\usage{
to_activitylog(eventlog)
}
\arguments{
\item{eventlog}{Object of class eventlog to be converted to activitylog}
}
\description{
Convert eventlog object to activitylog object.
}
