% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_eventlog.R
\name{to_eventlog}
\alias{to_eventlog}
\alias{to_eventlog.activitylog}
\title{Convert activitylog to eventlog}
\usage{
to_eventlog(activitylog)

\method{to_eventlog}{activitylog}(activitylog)
}
\arguments{
\item{activitylog}{Object of class activitylog to be converted to eventlog.}
}
\description{
Convert activitylog to eventlog
}
\section{Methods (by class)}{
\itemize{
\item \code{activitylog}: Convert activitylog to eventlog
}}

