% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_list.R
\name{trace_list}
\alias{trace_list}
\alias{trace_list.eventlog}
\alias{trace_list.activitylog}
\title{Trace list}
\usage{
trace_list(eventlog)

\method{trace_list}{eventlog}(eventlog)

\method{trace_list}{activitylog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog object}
}
\description{
Construct trace list
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Construct trace list for event log

\item \code{activitylog}: Construct trace list for activity log
}}

