% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{traces}
\alias{traces}
\alias{traces.log}
\alias{traces.grouped_log}
\title{Traces}
\usage{
traces(log, ..., eventlog = deprecated())

\method{traces}{log}(log, ..., eventlog = deprecated())

\method{traces}{grouped_log}(log, ..., eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{...}{Deprecated arguments}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
\code{traces} computes the different activity sequences of an event log
together with their absolute and relative frequencies.
Activity sequences are based on the start timestamp of activities.
}
\section{Methods (by class)}{
\itemize{
\item \code{traces(log)}: Construct traces list for eventlog

\item \code{traces(grouped_log)}: Construct list of traces for grouped log

}}
\seealso{
\code{\link{cases}}, \code{\link{eventlog}}
}
