% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup.R
\name{ungroup}
\alias{ungroup}
\alias{ungroup.activitylog}
\title{Ungroup log}
\usage{
ungroup(x, ...)

\method{ungroup}{activitylog}(x, ...)
}
\arguments{
\item{x}{Activitylog}

\item{...}{variables to remove from grouping}
}
\description{
Ungroup log
}
\section{Methods (by class)}{
\itemize{
\item \code{ungroup(activitylog)}: Ungroup columns in eventlog

}}
