% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup_eventlog.R
\name{ungroup_eventlog}
\alias{ungroup_eventlog}
\alias{ungroup_eventlog.eventlog}
\alias{ungroup_eventlog.grouped_log}
\title{Ungroup event log}
\usage{
ungroup_eventlog(log)

\method{ungroup_eventlog}{eventlog}(log)

\method{ungroup_eventlog}{grouped_log}(log)
}
\arguments{
\item{log}{Eventlog}
}
\description{
Remove groups from event log
}
\section{Methods (by class)}{
\itemize{
\item \code{ungroup_eventlog(eventlog)}: Remove groups from event log

\item \code{ungroup_eventlog(grouped_log)}: Remove groups from \code{\link{log}}.

}}
