% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_recode.R
\name{act_recode}
\alias{act_recode}
\alias{act_recode.log}
\alias{act_recode.grouped_log}
\title{Recode activity labels}
\usage{
act_recode(log, ..., eventlog = deprecated())

\method{act_recode}{log}(log, ..., eventlog = deprecated())

\method{act_recode}{grouped_log}(log, ..., eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{...}{A sequence of named character vectors of length one where the names gives the new label and the value gives the old label. Labels not mentioned will be left unchanged.}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Recode one or more activity labels through specifying their old and new label
}
\section{Methods (by class)}{
\itemize{
\item \code{act_recode(log)}: Recode activity labels of event log

\item \code{act_recode(grouped_log)}: Recode activity labels of event log

}}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}}, \code{\link{act_unite}}

Other Activity processing functions: 
\code{\link{act_collapse}()},
\code{\link{act_unite}()}
}
\concept{Activity processing functions}
