% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_recode.R
\name{act_recode}
\alias{act_recode}
\alias{act_recode.eventlog}
\title{Recode activity labels}
\usage{
act_recode(eventlog, ...)

\method{act_recode}{eventlog}(eventlog, ...)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}

\item{...}{A sequence of named character vectors of length one where the names gives the new label and the value gives the old label. Labels not mentioned will be left unchanged.}
}
\description{
Recode one or more activity labels through specifying their old and new label
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Recode activity labels of event log
}}

\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}}, \code{\link{act_unite}}

Other Activity processing functions: \code{\link{act_collapse}},
  \code{\link{act_unite}}
}
\concept{Activity processing functions}
