% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_unite.R
\name{act_unite}
\alias{act_unite}
\alias{act_unite.eventlog}
\title{Unite activity labels}
\usage{
act_unite(eventlog, ...)

\method{act_unite}{eventlog}(eventlog, ...)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}

\item{...}{A series of named character vectors. The activity labels in each vector will be replaced with the name.}
}
\description{
Recode two or different more activity labels two a uniform activity label
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Unite activity labels in event log
}}

\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}}, \code{\link{act_recode}}

Other Activity processing functions: \code{\link{act_collapse}},
  \code{\link{act_recode}}
}
\concept{Activity processing functions}
