% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{activities}
\alias{activities}
\title{Activities}
\usage{
activities(log, eventlog)
}
\arguments{
\item{log}{Object of class eventlog or activitylog.}

\item{eventlog}{Deprecated; please use log instead.}
}
\description{
Returns a \code{tibble}  containing a list of all activity types in the event log, with their absolute and relative frequency
}
\seealso{
\code{\link{activity_id}},\code{\link{activity_instance_id}}, \code{\link{eventlog}}
}
