% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{activities}
\alias{activities}
\title{Activities}
\usage{
activities(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns a \code{tbl_df}  containing a list of all activity types in the event log, with their absolute and relative frequency
}
\examples{

data(example_log)
activities(example_log)

}
\seealso{
\code{\link{activity_id}},\code{\link{activity_instance_id}}, \code{\link{eventlog}}
}

