% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities_to_eventlog.R
\name{activities_to_eventlog}
\alias{activities_to_eventlog}
\title{Create event log from list of activity instances}
\usage{
activities_to_eventlog(activity_log, case_id, activity_id, resource_id,
  timestamps)
}
\arguments{
\item{activity_log}{A data.frame where each row is an activity instances}

\item{case_id}{Column name of the case identifier}

\item{activity_id}{Column name of the activity identifier}

\item{resource_id}{Column name of the resource identifier}

\item{timestamps}{A vector of column names containing different timestamp. To column names will be transformed to lifecycle identifiers}
}
\description{
Create event log from list of activity instances
}
