% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_id.R
\name{activity_id}
\alias{activity_id}
\alias{activity_id.default}
\alias{activity_id.log_mapping}
\title{Activity classifier}
\usage{
activity_id(x)

\method{activity_id}{default}(x)

\method{activity_id}{log_mapping}(x)
}
\arguments{
\item{x}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}, or \code{\link{mapping}}.}
}
\description{
Get the activity classifier of an object of class \code{eventlog}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Retrieve activity identifier from log

\item \code{log_mapping}: Retrieve activity identifier from  mapping
}}

\seealso{
\code{\link{eventlog}}, \code{\link{activitylog}}, \code{\link{mapping}}

Other Classifiers: 
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
