% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_instance_id.R
\name{activity_instance_id}
\alias{activity_instance_id}
\title{Activity instance classifier}
\usage{
activity_instance_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the activity instance classifier of an object of class \code{eventlog}.
}
\examples{
data(example_log)
activity_instance_id(example_log)
}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}},
\code{\link{timestamp}},
\code{\link{case_id}}
}

