% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_labels.R
\name{activity_labels}
\alias{activity_labels}
\alias{activity_labels.log}
\title{Get vector of activity labels}
\usage{
activity_labels(eventlog)

\method{activity_labels}{log}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Retrieve a vector containing all unique activity labels
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Retrieve activity labels
}}

