% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_labels.R
\name{activity_labels}
\alias{activity_labels}
\alias{activity_labels.log}
\title{Get vector of activity labels}
\usage{
activity_labels(log, eventlog = deprecated())

\method{activity_labels}{log}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Retrieve a vector containing all unique activity labels
}
\section{Methods (by class)}{
\itemize{
\item \code{activity_labels(log)}: Retrieve activity labels

}}
