% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activitylog.R
\name{activitylog}
\alias{activitylog}
\title{Create activity log}
\usage{
activitylog(activitylog, case_id, activity_id, resource_id, lifecycle_ids)
}
\arguments{
\item{activitylog}{The data object to be used as activity log. This can be a
\code{data.frame} or \code{tbl_df}.}

\item{case_id}{The case classifier of the activity log. A character vector containing variable names of length 1 or more.}

\item{activity_id}{The activity classifier of the activity log. A character vector containing variable names of length 1 or more.}

\item{resource_id}{The resource identifier of the activity log. A character vector containing variable names of length 1 or more.}

\item{lifecycle_ids}{The columns with timestamps refering to different lifecycle events. A character vector of 1 or more.
These should have one of the folliwing names: "schedule","assign","reassign","start","suspend","resume","abort_activity","abort_case","complete","manualskip","autoskip".
These colomns should be of the Date or POSIXct class.}
}
\description{
Create activity log
}
