% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_end_activity.R
\name{add_end_activity}
\alias{add_end_activity}
\alias{add_start_activity}
\alias{add_end_activity.eventlog}
\alias{add_end_activity.grouped_eventlog}
\alias{add_start_activity.eventlog}
\alias{add_start_activity.grouped_eventlog}
\title{Add artificial start/end activities to}
\usage{
add_end_activity(eventlog, label)

add_start_activity(eventlog, label)

\method{add_end_activity}{eventlog}(eventlog, label = "End")

\method{add_end_activity}{grouped_eventlog}(eventlog, label = "End")

\method{add_start_activity}{eventlog}(eventlog, label = "Start")

\method{add_start_activity}{grouped_eventlog}(eventlog, label = "Start")
}
\arguments{
\item{eventlog}{Event log}

\item{label}{Start/end activity label}
}
\description{
Add artificial start/end activities to
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Add end activity to event log

\item \code{grouped_eventlog}: Add end activity to grouped event log

\item \code{eventlog}: Add start activity to event log

\item \code{grouped_eventlog}: Add start activity to grouped event log
}}

