% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange}
\alias{arrange}
\alias{arrange.grouped_eventlog}
\alias{arrange.activitylog}
\title{Arrange log}
\usage{
arrange(.data, ..., .by_group = FALSE)

\method{arrange}{grouped_eventlog}(.data, ...)

\method{arrange}{activitylog}(.data, ...)
}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{
  Arguments passed on to \code{\link[dplyr:arrange]{dplyr::arrange}}
  \describe{
    \item{\code{}}{}
  }}

\item{.by_group}{If \code{TRUE}, will sort first by grouping variable. Applies to
grouped data frames only.}
}
\description{
Arrange log
}
\section{Methods (by class)}{
\itemize{
\item \code{arrange(grouped_eventlog)}: Arrange an eventlog by group, maintaining all groups

\item \code{arrange(activitylog)}: Arrange an activitylog

}}
