% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\docType{import}
\name{arrange}
\alias{arrange}
\alias{arrange.grouped_eventlog}
\alias{arrange.activitylog}
\title{Arrange event log}
\usage{
\method{arrange}{grouped_eventlog}(.data, ...)

\method{arrange}{activitylog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Variables to arrange on}
}
\section{Methods (by class)}{
\itemize{
\item \code{grouped_eventlog}: Arrange an eventlog by group, maintaining all groups

\item \code{activitylog}: Arrange an activitylog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}}
}}

