% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupaR-package.R, R/bupar.R
\docType{package}
\name{bupaR}
\alias{bupaR}
\alias{bupaR-package}
\title{bupaR - Business Process Analysis in R}
\description{
Functionalities for process analysis in R. This packages implements an S3-class for event log objects, and related handler functions. Imports related packages for subsetting event data, computation of descriptive statistics, handling of Petri Net objects and visualization of process maps.

Functionalities for process analysis in R. This packages implements an S3-class for event log objects, and related handler functions. Imports related packages for subsetting event data, computation of descriptive statistics, handling of Petri Net objects and visualization of process maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.bupar.net}
  \item \url{https://github.com/bupaverse/bupaR}
}

}
\author{
\strong{Maintainer}: Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be}

Other contributors:
\itemize{
  \item Felix Mannhardt \email{Felix.Mannhardt@sintef.no} [contributor]
  \item Niels Martin \email{niels.martin@uhasselt.be} [contributor]
  \item Greg Van Houdt \email{greg.vanhoudt@uhasselt.be} [contributor]
}

}
\keyword{internal}
