% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bupar.R
\docType{package}
\name{bupaR}
\alias{bupaR}
\alias{bupaR-package}
\title{bupaR - Business Process Analysis in R}
\description{
bupaR is a collection of functions and other packages which can be used for (business) process analysis in R.
}

