% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_id.r
\name{case_id}
\alias{case_id}
\title{Case classifier}
\usage{
case_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the case classifier of an object of class \code{eventlog}
}
\examples{


data(example_log)
case_id(example_log)

}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_id}},
\code{\link{lifecycle_id}}, \code{\link{activity_instance_id}}
}

