% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_labels.R
\name{case_labels}
\alias{case_labels}
\alias{case_labels.log}
\title{Case labels}
\usage{
case_labels(log, eventlog = deprecated())

\method{case_labels}{log}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Retrieve a vector containing all unique case labels
}
\section{Methods (by class)}{
\itemize{
\item \code{case_labels(log)}: Retrieve case labels from log

}}
