% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_list.R
\name{case_list}
\alias{case_list}
\alias{case_list.eventlog}
\alias{case_list.activitylog}
\title{Case list}
\usage{
case_list(log, .keep_trace_list)

\method{case_list}{eventlog}(log, .keep_trace_list = FALSE)

\method{case_list}{activitylog}(log, .keep_trace_list = FALSE)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{.keep_trace_list}{Logical (default is \code{FALSE}): If \code{TRUE}, keeps the trace as a \code{list}.
If \code{FALSE}, only the concatenated string representation of the trace is kept.}
}
\description{
Construct list of cases
}
\section{Methods (by class)}{
\itemize{
\item \code{case_list(eventlog)}: Return case list

\item \code{case_list(activitylog)}: Return case list

}}
