% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_list.R
\name{case_list}
\alias{case_list}
\alias{case_list.eventlog}
\title{Case list}
\usage{
case_list(log, eventlog, .keep_trace_list)

\method{case_list}{eventlog}(log, eventlog = deprecated(), .keep_trace_list = FALSE)
}
\arguments{
\item{eventlog}{Eventlog object}
}
\description{
Construct list of cases
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Return case list
}}

