% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_list.R
\name{case_list}
\alias{case_list}
\alias{case_list.eventlog}
\title{Case list}
\usage{
case_list(eventlog)

\method{case_list}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog object}
}
\description{
Construct list of cases
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Return case list
}}

