% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.r
\name{cases}
\alias{cases}
\alias{cases.log}
\alias{cases.eventlog}
\alias{cases.activitylog}
\title{Cases}
\usage{
cases(log, eventlog = deprecated(), ...)

\method{cases}{log}(log, eventlog = deprecated(), ...)

\method{cases}{eventlog}(log, eventlog = deprecated(), ...)

\method{cases}{activitylog}(log, eventlog = deprecated(), ...)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}

\item{...}{Other (optional) arguments passed on to methods. See \code{\link{durations}} for more options.}
}
\description{
Provides a fine-grained summary of an event log with characteristics for each case: the number of events,
the number of activity types, the timespan, the trace, the duration, and the first and last event type.
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Constructy list of cases in a log

\item \code{eventlog}: Constructy list of cases in an eventlog

\item \code{activitylog}: Constructy list of cases in an activitylog
}}

\seealso{
\code{\link{case_list}},\code{\link{durations}}
}
