% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.r
\name{cases}
\alias{cases}
\title{Cases}
\usage{
cases(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Provides a fine-grained summary of an event log with characteristics for each case:  the number of events,
the number of activity types, the timespan, the trace, the duration and the first and last event type.
}
\examples{


data(example_log)
cases(example_log)

}

