% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\alias{count.log}
\alias{count.grouped_log}
\title{Count log}
\usage{
count(x, ..., wt = NULL, sort = FALSE, name = NULL)

\method{count}{log}(x, ...)

\method{count}{grouped_log}(x, ...)
}
\arguments{
\item{x}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Additional arguments passed to \link[=count]{dplyr}}

\item{wt}{<\code{\link[rlang:args_data_masking]{data-masking}}> Frequency weights.
Can be \code{NULL} or a variable:
\itemize{
\item If \code{NULL} (the default), counts the number of rows in each group.
\item If a variable, computes \code{sum(wt)} for each group.
}}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}
}
\description{
Count log
}
\section{Methods (by class)}{
\itemize{
\item \code{count(log)}: Count log

\item \code{count(grouped_log)}: Count grouped log

}}
