% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durations.R
\name{durations}
\alias{durations}
\alias{durations.eventlog}
\title{Durations}
\usage{
durations(eventlog, units)

\method{durations}{eventlog}(eventlog, units = "days")
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{units}{The time unit in which the throughput times should be reported.}
}
\description{
Computes the throughput times of each case.
Throughput time is defined as the interval between the start of the first event and the completion of the last event.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Compute durations from eventlog
}}

