% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\docType{import}
\name{fill}
\alias{fill}
\title{Fill event log}
\arguments{
\item{data}{Eventlog}

\item{...}{options for fill}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyr}{\code{\link[tidyr]{fill}}}
}}

