% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.eventlog.R
\docType{import}
\name{fill}
\alias{fill}
\alias{fill.eventlog}
\alias{fill.grouped_eventlog}
\title{Fill event log}
\usage{
\method{fill}{eventlog}(data, ..., .direction)

\method{fill}{grouped_eventlog}(data, ..., .direction)
}
\arguments{
\item{...}{options for fill}

\item{.data}{Eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Fill eventlog

\item \code{grouped_eventlog}: Fill grouped eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyr}{\code{\link[tidyr]{fill}}}
}}

