% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{import}
\name{filter}
\alias{filter}
\title{Filter event log}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Conditions to filter on}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}
}}

