% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_attributes.R
\name{filter_attributes}
\alias{filter_attributes}
\alias{filter_attributes.eventlog}
\alias{filter_attributes.grouped_eventlog}
\title{Generic filter function for eventlog}
\usage{
filter_attributes(eventlog, ...)

\method{filter_attributes}{eventlog}(eventlog, ...)

\method{filter_attributes}{grouped_eventlog}(eventlog, ...)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{...}{Filter conditions}
}
\description{
Generic filter function for eventlog
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter eventlog using attributes

\item \code{grouped_eventlog}: Filter grouped eventlog using attributes
}}

