% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n.R
\name{first_n}
\alias{first_n}
\alias{first_n.eventlog}
\alias{first_n.grouped_eventlog}
\alias{first_n.activitylog}
\alias{first_n.grouped_activitylog}
\title{first_n}
\usage{
first_n(log, eventlog = deprecated(), n)

\method{first_n}{eventlog}(log, eventlog = deprecated(), n)

\method{first_n}{grouped_eventlog}(log, eventlog = deprecated(), n)

\method{first_n}{activitylog}(log, eventlog = deprecated(), n)

\method{first_n}{grouped_activitylog}(log, eventlog = deprecated(), n)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}

\item{n}{\code{\link{integer}}: The number of activity instances to select.}
}
\description{
Select first n activity instances.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Select first n activity instances of an \code{\link{eventlog}}.

\item \code{grouped_eventlog}: Select first n activity instances of a \code{\link{grouped_eventlog}}.

\item \code{activitylog}: Select first n activity instances of an \code{\link{activitylog}}.

\item \code{grouped_activitylog}: Select first n activity instances of a \code{\link{grouped_activitylog}}.
}}

