% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n.R
\name{first_n}
\alias{first_n}
\alias{first_n.eventlog}
\alias{first_n.grouped_eventlog}
\title{Select first n activity instances}
\usage{
first_n(eventlog, n)

\method{first_n}{eventlog}(eventlog, n)

\method{first_n}{grouped_eventlog}(eventlog, n)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{n}{Integer value}
}
\description{
Select first n activity instances
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Select first n activity instances in event log

\item \code{grouped_eventlog}: Select first n activity instances  in grouped event log
}}

