% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\docType{import}
\name{group_by}
\alias{group_by}
\title{Group event log}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Variables to group by}

\item{.add}{Add grouping variables to existing ones}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}}
}}

