% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R
\docType{import}
\name{group_by}
\alias{group_by}
\alias{group_by.eventlog}
\title{Group event log}
\usage{
\method{group_by}{eventlog}(.data, ..., add = F)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Variables to group by}

\item{add}{Add grouping variables to existing ones}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}}
}}

