% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R
\name{group_by_activity}
\alias{group_by_activity}
\alias{group_by_activity.eventlog}
\title{Group event log on activity id}
\usage{
group_by_activity(eventlog)

\method{group_by_activity}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Group an event log by activity identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on activity identifier
}}

