% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_activity_instance.R
\name{group_by_activity_instance}
\alias{group_by_activity_instance}
\title{Group event log on activity instance id}
\usage{
group_by_activity_instance(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}
}
\description{
Group an event log by activity instance identifier
}
