% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R,
%   R/group_by_activity_instance.R
\name{group_by_activity_instance}
\alias{group_by_activity_instance}
\alias{group_by_activity_instance.eventlog}
\title{Group event log on activity instance id}
\usage{
group_by_activity_instance(log)

\method{group_by_activity_instance}{eventlog}(log)

group_by_activity_instance(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}

\item{eventlog}{Eventlog}
}
\description{
Group an event log by activity instance identifier

Group an event log by activity instance identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on activity instance identifier
}}

