% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R
\name{group_by_activity_instance}
\alias{group_by_activity_instance}
\alias{group_by_activity_instance.eventlog}
\title{Group event log on activity instance id}
\usage{
group_by_activity_instance(eventlog)

\method{group_by_activity_instance}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Group an event log by activity instance identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on activity instance identifier
}}

