% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R, R/group_by_case.R
\name{group_by_case}
\alias{group_by_case}
\alias{group_by_case.eventlog}
\title{Group event log on case id}
\usage{
group_by_case(log)

\method{group_by_case}{eventlog}(log)

group_by_case(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}

\item{eventlog}{Eventlog}
}
\description{
Group an event log by case identifier

Group an event log by case identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on case identifier
}}

