% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R
\name{group_by_resource}
\alias{group_by_resource}
\alias{group_by_resource.eventlog}
\title{Group event log on resource id}
\usage{
group_by_resource(eventlog)

\method{group_by_resource}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Group an event log by resource identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Group eventlog on resource identifier
}}

