% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_resource_activity.R
\name{group_by_resource_activity}
\alias{group_by_resource_activity}
\title{Group event log on resource and activity id}
\usage{
group_by_resource_activity(log)
}
\arguments{
\item{log}{Object of classes eventlog or activitylog}
}
\description{
Group an event log by resource and activity identifier
}
