% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_n.R
\name{last_n}
\alias{last_n}
\alias{last_n.eventlog}
\alias{last_n.activitylog}
\alias{last_n.grouped_log}
\title{last_n}
\usage{
last_n(log, n, eventlog = deprecated())

\method{last_n}{eventlog}(log, n, eventlog = deprecated())

\method{last_n}{activitylog}(log, n, eventlog = deprecated())

\method{last_n}{grouped_log}(log, n, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{n}{\code{\link{integer}}: The number of activity instances to select.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Select last n activity instances
}
\section{Methods (by class)}{
\itemize{
\item \code{last_n(eventlog)}: Select last n activity instances of an \code{\link{eventlog}}.

\item \code{last_n(activitylog)}: Select last n activity instances of an \code{\link{activitylog}}.

\item \code{last_n(grouped_log)}: Select last n activity instances of a \code{\link{grouped_log}}.

}}
