% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_id.R
\name{lifecycle_id}
\alias{lifecycle_id}
\alias{lifecycle_id.eventlog}
\alias{lifecycle_id.eventlog_mapping}
\alias{lifecycle_id.activitylog}
\alias{lifecycle_id.activitylog_mapping}
\title{Life cycle classifier}
\usage{
lifecycle_id(x)

\method{lifecycle_id}{eventlog}(x)

\method{lifecycle_id}{eventlog_mapping}(x)

\method{lifecycle_id}{activitylog}(x)

\method{lifecycle_id}{activitylog_mapping}(x)
}
\arguments{
\item{x}{An \code{eventlog} of \code{eventlog_mapping}}
}
\description{
Get the life_cycle_id of an object of class \code{eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve lifecycle identifier from eventlog

\item \code{eventlog_mapping}: Retrieve lifecycle identifier from eventlog mapping

\item \code{activitylog}: Retrieve lifecycle identifier from activitylog

\item \code{activitylog_mapping}: Retrieve lifecycle identifier from activitylog mapping
}}

\seealso{
Other Eventlog classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_ids}()},
\code{\link{mapping}()},
\code{\link{resource_id}()},
\code{\link{timestamp}()}
}
\concept{Eventlog classifiers}
