% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_id.R
\name{lifecycle_id}
\alias{lifecycle_id}
\title{Life cycle classifier}
\usage{
lifecycle_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the life_cycle_id of an object of class \code{eventlog}
}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_instance_id}}
}

