% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_labels.R
\name{lifecycle_labels}
\alias{lifecycle_labels}
\alias{lifecycle_labels.eventlog}
\alias{lifecycle_labels.activitylog}
\title{Get vector of lifecycle labels.}
\usage{
lifecycle_labels(log)

\method{lifecycle_labels}{eventlog}(log)

\method{lifecycle_labels}{activitylog}(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Retrieve a vector containing all unique lifecycle labels.
}
\section{Methods (by class)}{
\itemize{
\item \code{lifecycle_labels(eventlog)}: Retrieve lifecycle labels from an \code{\link{eventlog}}.

\item \code{lifecycle_labels(activitylog)}: Retrieve lifecycle labels from an \code{\link{activitylog}}.

}}
\seealso{
\code{\link{lifecycle_id}}
}
