% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle_labels.R
\name{lifecycle_labels}
\alias{lifecycle_labels}
\alias{lifecycle_labels.eventlog}
\alias{lifecycle_labels.activitylog}
\title{Get vector of lifecycle labels.}
\usage{
lifecycle_labels(log, eventlog = deprecated())

\method{lifecycle_labels}{eventlog}(log, eventlog = deprecated())

\method{lifecycle_labels}{activitylog}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Retrieve a vector containing all unique lifecycle labels.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve lifecycle labels from an \code{\link{eventlog}}.

\item \code{activitylog}: Retrieve lifecycle labels from an \code{\link{activitylog}}.
}}

\seealso{
\code{\link{lifecycle_id}},\code{\link{lifecycle_ids}}
}
