% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycles.R
\name{lifecycles}
\alias{lifecycles}
\alias{lifecycles.eventlog}
\alias{lifecycles.grouped_eventlog}
\title{Life cycles}
\usage{
lifecycles(eventlog)

\method{lifecycles}{eventlog}(eventlog)

\method{lifecycles}{grouped_eventlog}(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns a \code{tibble}  containing a list of all life cycle types in the event log, with their absolute and relative frequency (# events)
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Generate lifecycle list for eventlog

\item \code{grouped_eventlog}: Generate lifecycle list for grouped eventlog
}}

\seealso{
\code{\link{lifecycle_id}}, \code{\link{eventlog}}
}
