% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\alias{mapping.eventlog}
\alias{mapping.activitylog}
\title{Mapping}
\usage{
mapping(log, eventlog = deprecated())

\method{mapping}{eventlog}(log, eventlog = deprecated())

\method{mapping}{activitylog}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Prints the mapping of an event log object.
}
\section{Methods (by class)}{
\itemize{
\item \code{mapping(eventlog)}: Retrieve identifier mapping from eventlog

\item \code{mapping(activitylog)}: Retrieve identifier mapping from activitylog

}}
\concept{Eventlog classifiers}
