% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\docType{import}
\name{mutate}
\alias{mutate}
\title{Mutate event log}
\arguments{
\item{.data}{Eventlog}

\item{...}{New variables}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}
}}

