% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.eventlog.R
\docType{import}
\name{mutate}
\alias{mutate}
\alias{mutate.eventlog}
\alias{mutate.grouped_eventlog}
\title{Mutate event log}
\usage{
\method{mutate}{eventlog}(.data, ...)

\method{mutate}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{New variables}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Mutate eventlog

\item \code{grouped_eventlog}: Mutate grouped eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}
}}

