% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_activities.R
\name{n_activities}
\alias{n_activities}
\title{n_activities}
\usage{
n_activities(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of activities in an event log
}

