% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_activity_instances.R
\name{n_activity_instances}
\alias{n_activity_instances}
\alias{n_activity_instances.eventlog}
\alias{n_activity_instances.grouped_eventlog}
\alias{n_activity_instances.activitylog}
\alias{n_activity_instances.grouped_activitylog}
\title{n_activity_instances}
\usage{
n_activity_instances(log)

\method{n_activity_instances}{eventlog}(log)

\method{n_activity_instances}{grouped_eventlog}(log)

\method{n_activity_instances}{activitylog}(log)

\method{n_activity_instances}{grouped_activitylog}(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Returns the number of activity instances in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{n_activity_instances(eventlog)}: eventlog

\item \code{n_activity_instances(grouped_eventlog)}: grouped_eventlog

\item \code{n_activity_instances(activitylog)}: activitylog

\item \code{n_activity_instances(grouped_activitylog)}: grouped_activitylog

}}
\seealso{
Other Counters: 
\code{\link{n_activities}()},
\code{\link{n_cases}()},
\code{\link{n_events}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Counters}
