% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_activity_instances.R
\name{n_activity_instances}
\alias{n_activity_instances}
\title{n_activity_instances}
\usage{
n_activity_instances(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of activity instances in an event log
}

