% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cases.R
\name{n_cases}
\alias{n_cases}
\title{n_cases}
\usage{
n_cases(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of cases in an event log
}

