% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_events.R
\name{n_events}
\alias{n_events}
\alias{n_events.eventlog}
\alias{n_events.grouped_eventlog}
\alias{n_events.activitylog}
\alias{n_events.grouped_activitylog}
\title{n_events}
\usage{
n_events(log)

\method{n_events}{eventlog}(log)

\method{n_events}{grouped_eventlog}(log)

\method{n_events}{activitylog}(log)

\method{n_events}{grouped_activitylog}(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
Returns the number of events in an event log.
}
\section{Methods (by class)}{
\itemize{
\item \code{n_events(eventlog)}: Count number of events in an \code{\link{eventlog}}.

\item \code{n_events(grouped_eventlog)}: Count number of events in a \code{\link{grouped_eventlog}}.

\item \code{n_events(activitylog)}: Count number of events in an \code{\link{activitylog}}.

\item \code{n_events(grouped_activitylog)}: Count number of events in an \code{\link{grouped_activitylog}}.

}}
\seealso{
Other Counters: 
\code{\link{n_activities}()},
\code{\link{n_activity_instances}()},
\code{\link{n_cases}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Counters}
