% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_events.R
\name{n_events}
\alias{n_events}
\alias{n_events.eventlog}
\alias{n_events.grouped_eventlog}
\alias{n_events.activitylog}
\title{n_events}
\usage{
n_events(log, eventlog = deprecated())

\method{n_events}{eventlog}(log, eventlog = deprecated())

\method{n_events}{grouped_eventlog}(log, eventlog = deprecated())

\method{n_events}{activitylog}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}}, \code{\link{eventlog}}, or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Returns the number of events in an event log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Count number of resources in eventlog

\item \code{grouped_eventlog}: Count number of resource in eventlog

\item \code{activitylog}: Count number of resources in eventlog
}}

\seealso{
Other Eventlog count functions: 
\code{\link{n_activities}()},
\code{\link{n_activity_instances}()},
\code{\link{n_cases}()},
\code{\link{n_resources}()},
\code{\link{n_traces}()}
}
\concept{Eventlog count functions}
