% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_events.R
\name{n_events}
\alias{n_events}
\title{n_events}
\usage{
n_events(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of events in an event log
}

