% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_resources.R
\name{n_resources}
\alias{n_resources}
\title{n_resources}
\usage{
n_resources(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of resources in an event log
}

