% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_traces.R
\name{n_traces}
\alias{n_traces}
\title{n_traces}
\usage{
n_traces(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of traces in an event log
}

