% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_map.R
\name{re_map}
\alias{re_map}
\title{Re map}
\usage{
re_map(x, mapping)
}
\arguments{
\item{x}{The eventlog/activitylog data to be used.}

\item{mapping}{An existing eventlog mapping created by the mapping function}
}
\description{
Construct an eventlog using an existing mapping.
}
