% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{ymd_hms}
\alias{ymd_hm}
\alias{ymd_h}
\alias{ymd}
\alias{dmy_hms}
\alias{dmy_hm}
\alias{dmy_h}
\alias{dmy}
\alias{mdy_hms}
\alias{mdy_hm}
\alias{mdy_h}
\alias{mdy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate:ymd]{dmy}}, \code{\link[lubridate:ymd_hms]{dmy_h}}, \code{\link[lubridate:ymd_hms]{dmy_hm}}, \code{\link[lubridate:ymd_hms]{dmy_hms}}, \code{\link[lubridate:ymd]{mdy}}, \code{\link[lubridate:ymd_hms]{mdy_h}}, \code{\link[lubridate:ymd_hms]{mdy_hm}}, \code{\link[lubridate:ymd_hms]{mdy_hms}}, \code{\link[lubridate]{ymd}}, \code{\link[lubridate:ymd_hms]{ymd_h}}, \code{\link[lubridate:ymd_hms]{ymd_hm}}, \code{\link[lubridate]{ymd_hms}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

