% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\docType{import}
\name{rename}
\alias{rename}
\alias{rename.log}
\alias{rename.grouped_log}
\title{Rename log}
\usage{
\method{rename}{log}(.data, ...)

\method{rename}{grouped_log}(.data, ...)
}
\arguments{
\item{.data}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Variables to rename. Use "new_name" = "old_name" to rename selected variables.}
}
\section{Methods (by class)}{
\itemize{
\item \code{rename(log)}: Rename log

\item \code{rename(grouped_log)}: Rename grouped log

}}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{rename}}}
}}

