% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_id.R
\name{resource_id}
\alias{resource_id}
\title{Resource classifier}
\usage{
resource_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the resource classifier of an object of class \code{eventlog}.
}
\examples{
data(example_log)
resource_id(example_log)
}
\seealso{
\code{\link{eventlog}}, \code{\link{case_id}}, \code{\link{activity_instance_id}},
\code{\link{timestamp}}
}

