% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_labels.R
\name{resource_labels}
\alias{resource_labels}
\alias{resource_labels.default}
\title{Get vector of resource labels}
\usage{
resource_labels(log, eventlog, ...)

\method{resource_labels}{default}(log, eventlog, ...)
}
\arguments{
\item{log}{Object of class eventlog or activitylog.}

\item{eventlog}{Deprecated; please use log instead.}

\item{...}{Unused.}
}
\description{
Retrieve a vector containing all unique resource labels
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Retrieve resource labels from eventlog
}}

