% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resource_labels}
\alias{resource_labels}
\alias{resource_labels.eventlog}
\title{Get vector of resource labels}
\usage{
resource_labels(eventlog)

\method{resource_labels}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Retrieve a vector containing all unique resource labels
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve resource labels from eventlog
}}

