% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\alias{resources.eventlog}
\alias{resources.activitylog}
\alias{resources.grouped_log}
\title{Resources}
\usage{
resources(log, eventlog = deprecated(), ...)

\method{resources}{eventlog}(log, eventlog = deprecated(), ...)

\method{resources}{activitylog}(log, eventlog = deprecated(), ...)

\method{resources}{grouped_log}(log, eventlog = deprecated(), ...)
}
\arguments{
\item{log}{Object of class eventlog or activitylog.}

\item{eventlog}{Deprecated; please use log instead.}

\item{...}{Unused.}
}
\description{
Returns a \code{tibble}  containing a list of all resources in the event log, with there absolute and relative frequency
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Generate resource list for eventlog

\item \code{activitylog}: Generate resource list for activitylog

\item \code{grouped_log}: Compute activity frequencies
}}

\seealso{
\code{\link{resource_id}}, \code{\link{eventlog}}
}
