% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\alias{resources.eventlog}
\alias{resources.activitylog}
\alias{resources.grouped_log}
\title{Resources}
\usage{
resources(log, eventlog = deprecated())

\method{resources}{eventlog}(log, eventlog = deprecated())

\method{resources}{activitylog}(log, eventlog = deprecated())

\method{resources}{grouped_log}(log, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Returns a \code{tibble}  containing a list of all resources in the event log, with there absolute and relative frequency
}
\section{Methods (by class)}{
\itemize{
\item \code{resources(eventlog)}: Generate resource list for eventlog

\item \code{resources(activitylog)}: Generate resource list for activitylog

\item \code{resources(grouped_log)}: Compute activity frequencies

}}
\seealso{
\code{\link{resource_id}}, \code{\link{eventlog}}
}
