% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\title{Resources}
\usage{
resources(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns a \code{tbl_df}  containing a list of all resources in the event log, with there absolute and relative frequency
}
\seealso{
\code{\link{resource_id}}, \code{\link{eventlog}}
}

