% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_n.R
\name{sample_n}
\alias{sample_n}
\alias{sample_n.eventlog}
\alias{sample_n.grouped_eventlog}
\title{Sample function for eventlog}
\usage{
sample_n(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

\method{sample_n}{eventlog}(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)

\method{sample_n}{grouped_eventlog}(tbl, size, replace = FALSE, weight = NULL, .env = NULL, ...)
}
\arguments{
\item{tbl}{Deprecated; please use \code{log} instead.}

\item{size}{\code{\link{integer}}: Number of cases to sample}

\item{replace}{\code{\link{logical}} (default \code{FALSE}): Sample with replacement \code{TRUE} or without \code{FALSE}.}

\item{weight}{Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input.
Weights are automatically standardised to sum to 1.}

\item{.env}{Deprecated; please don't use.}

\item{...}{ignored}
}
\description{
Sample function for eventlog
}
\section{Methods (by class)}{
\itemize{
\item \code{sample_n(eventlog)}: Sample n cases of eventlog

\item \code{sample_n(grouped_eventlog)}: Stratified sampling of a grouped eventlog: sample n cases within each group

}}
\seealso{
\code{\link{slice_sample}}
}
