% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\docType{import}
\name{select}
\alias{select}
\title{Select event log}
\arguments{
\item{force_df}{If TRUE, result will no longer be a event log when not all id columns are selected.}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{select}}}
}}

