% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_activity_id.R
\name{set_activity_id}
\alias{set_activity_id}
\alias{set_activity_id.default}
\title{Set activity id of log}
\usage{
set_activity_id(log, activity_id, eventlog = deprecated())

\method{set_activity_id}{default}(log, activity_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{activity_id}{New activity id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set activity id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Set activity id
}}

\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
