% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_activity_instance_id.R
\name{set_activity_instance_id}
\alias{set_activity_instance_id}
\alias{set_activity_instance_id.eventlog}
\title{Set activity instance id of log}
\usage{
set_activity_instance_id(log, activity_instance_id, eventlog = deprecated())

\method{set_activity_instance_id}{eventlog}(log, activity_instance_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}}.}

\item{activity_instance_id}{New activity_instance id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set activity instance id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Set activity_instance_id of eventlog
}}

