% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_case_id.R
\name{set_case_id}
\alias{set_case_id}
\alias{set_case_id.default}
\title{Set case id of log}
\usage{
set_case_id(log, case_id)

\method{set_case_id}{default}(log, case_id)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{case_id}{New case id}
}
\description{
Set case id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{set_case_id(default)}: Set case id

}}
\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
