% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_eventlog_attributes.R
\name{set_case_id}
\alias{set_case_id}
\alias{set_activity_id}
\alias{set_activity_instance_id}
\alias{set_timestamp}
\alias{set_resource_id}
\alias{set_lifecycle_id}
\alias{set_case_id.eventlog}
\alias{set_case_id.grouped_eventlog}
\alias{set_activity_id.eventlog}
\alias{set_activity_id.grouped_eventlog}
\alias{set_activity_instance_id.eventlog}
\alias{set_activity_instance_id.grouped_eventlog}
\alias{set_timestamp.eventlog}
\alias{set_timestamp.grouped_eventlog}
\alias{set_resource_id.eventlog}
\alias{set_resource_id.grouped_eventlog}
\alias{set_lifecycle_id.eventlog}
\alias{set_lifecycle_id.grouped_eventlog}
\title{Set individual attributes of event log}
\usage{
set_case_id(eventlog, case_id)

set_activity_id(eventlog, activity_id)

set_activity_instance_id(eventlog, activity_instance_id)

set_timestamp(eventlog, timestamp)

set_resource_id(eventlog, resource_id)

set_lifecycle_id(eventlog, lifecycle_id)

\method{set_case_id}{eventlog}(eventlog, case_id)

\method{set_case_id}{grouped_eventlog}(eventlog, case_id)

\method{set_activity_id}{eventlog}(eventlog, activity_id)

\method{set_activity_id}{grouped_eventlog}(eventlog, activity_id)

\method{set_activity_instance_id}{eventlog}(eventlog, activity_instance_id)

\method{set_activity_instance_id}{grouped_eventlog}(eventlog,
  activity_instance_id)

\method{set_timestamp}{eventlog}(eventlog, timestamp)

\method{set_timestamp}{grouped_eventlog}(eventlog, timestamp)

\method{set_resource_id}{eventlog}(eventlog, resource_id)

\method{set_resource_id}{grouped_eventlog}(eventlog, resource_id)

\method{set_lifecycle_id}{eventlog}(eventlog, lifecycle_id)

\method{set_lifecycle_id}{grouped_eventlog}(eventlog, lifecycle_id)
}
\arguments{
\item{eventlog}{Event log object}

\item{case_id}{New case id}

\item{activity_id}{New activity id}

\item{activity_instance_id}{New activity instance id}

\item{timestamp}{New timestamp}

\item{resource_id}{New resource id}

\item{lifecycle_id}{New lifecycle id}
}
\description{
Set individual attributes of event log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Set case id of event log

\item \code{grouped_eventlog}: Set case id of grouped event log

\item \code{eventlog}: Set activity id of event log

\item \code{grouped_eventlog}: Set activity id of grouped event log

\item \code{eventlog}: Set activity instance id of event log

\item \code{grouped_eventlog}: Set activity instance id of grouped event log

\item \code{eventlog}: Set timestamp of event log

\item \code{grouped_eventlog}: Set timestamp of grouped event log

\item \code{eventlog}: Set resource_id of event log

\item \code{grouped_eventlog}: Set resource_id of grouped event log

\item \code{eventlog}: Set lifecycle_id of event log

\item \code{grouped_eventlog}: Set lifecycle_id of grouped event log
}}

