% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lifecycle_id.R
\name{set_lifecycle_id}
\alias{set_lifecycle_id}
\alias{set_lifecycle_id.default}
\title{Set lifecycle id of log}
\usage{
set_lifecycle_id(log, lifecycle_id)

\method{set_lifecycle_id}{default}(log, lifecycle_id)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{lifecycle_id}{New lifecycle id. Can be multiple in case of activitylog}
}
\description{
Set lifecycle id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{set_lifecycle_id(default)}: Set lifecycle id

}}
\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
