% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_resource_id.R
\name{set_resource_id}
\alias{set_resource_id}
\alias{set_resource_id.default}
\title{Set resource id of log}
\usage{
set_resource_id(log, resource_id, eventlog = deprecated())

\method{set_resource_id}{default}(log, resource_id, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{resource_id}{New resource id}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set resource id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Set resource id
}}

\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamps}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
