% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_resource_id.R
\name{set_resource_id}
\alias{set_resource_id}
\alias{set_resource_id.default}
\title{Set resource id of log}
\usage{
set_resource_id(log, resource_id)

\method{set_resource_id}{default}(log, resource_id)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{resource_id}{New resource id}
}
\description{
Set resource id of log
}
\section{Methods (by class)}{
\itemize{
\item \code{set_resource_id(default)}: Set resource id

}}
\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
