% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_timestamp.R
\name{set_timestamp}
\alias{set_timestamp}
\alias{set_timestamp.eventlog}
\title{Set timestamp of eventlog}
\usage{
set_timestamp(log, timestamp)

\method{set_timestamp}{eventlog}(log, timestamp)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{timestamp}{New timestamp}
}
\description{
Set timestamp of eventlog
}
\section{Methods (by class)}{
\itemize{
\item \code{set_timestamp(eventlog)}: Set timestamp of eventlog

}}
\seealso{
Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{timestamp}()},
\code{\link{timestamps}()}
}
\concept{Classifiers}
