% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_timestamp.R
\name{set_timestamp}
\alias{set_timestamp}
\alias{set_timestamp.eventlog}
\title{Set timestamp of eventlog}
\usage{
set_timestamp(log, timestamp, eventlog = deprecated())

\method{set_timestamp}{eventlog}(log, timestamp, eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}}.}

\item{timestamp}{New timestamp}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Set timestamp of eventlog
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Set timestamp of eventlog
}}

