% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R, R/slice_activities.R, R/slice_events.R
\docType{import}
\name{slice}
\alias{slice}
\alias{slice.log}
\alias{slice.grouped_log}
\alias{slice_activities.eventlog}
\alias{slice_activities.activitylog}
\alias{slice_activities.grouped_log}
\alias{slice_events.eventlog}
\alias{slice_events.grouped_eventlog}
\title{Slice function for event log}
\usage{
\method{slice}{log}(.data, ...)

\method{slice}{grouped_log}(.data, ...)

\method{slice_activities}{eventlog}(.data, ...)

\method{slice_activities}{activitylog}(.data, ...)

\method{slice_activities}{grouped_log}(.data, ...)

\method{slice_events}{eventlog}(.data, ...)

\method{slice_events}{grouped_eventlog}(.data, ...)
}
\section{Methods (by class)}{
\itemize{
\item \code{slice(log)}: Slice n cases of a log

\item \code{slice(grouped_log)}: Slice grouped log: take slice of cases from each group.

}}
\section{Functions}{
\itemize{
\item \code{slice_activities(eventlog)}: Take a slice of activity instances from event log

\item \code{slice_activities(activitylog)}: Take a slice of activity instances from activity log

\item \code{slice_activities(grouped_log)}: Take a slice of activity instances from grouped event log

\item \code{slice_events(eventlog)}: Take a slice of events from event log

\item \code{slice_events(grouped_eventlog)}: Take a slice of events from grouped event log

}}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{slice}}}
}}

