% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R, R/slice_activities.R, R/slice_events.R
\docType{import}
\name{slice}
\alias{slice}
\alias{slice.eventlog}
\alias{slice.grouped_eventlog}
\alias{slice_activities.eventlog}
\alias{slice_activities.grouped_eventlog}
\alias{slice_events.eventlog}
\alias{slice_events.grouped_eventlog}
\title{Slice function for event log}
\usage{
\method{slice}{eventlog}(.data, ...)

\method{slice}{grouped_eventlog}(.data, ...)

\method{slice_activities}{eventlog}(.data, ...)

\method{slice_activities}{grouped_eventlog}(.data, ...)

\method{slice_events}{eventlog}(.data, ...)

\method{slice_events}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Additional Arguments}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Slice n cases of an eventlog

\item \code{grouped_eventlog}: Slice grouped eventlog: take slice of cases from each group.

\item \code{eventlog}: Take a slice of activity instances from event log

\item \code{grouped_eventlog}: Take a slice of activity instances from grouped event log

\item \code{eventlog}: Take a slice of events from event log

\item \code{grouped_eventlog}: Take a slice of events from grouped event log
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{slice}}}
}}

