% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.eventlog.R
\docType{import}
\name{slice}
\alias{slice}
\alias{slice.eventlog}
\alias{slice.grouped_eventlog}
\title{Slice function for event log}
\usage{
\method{slice}{eventlog}(.data, ...)

\method{slice}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Additional Arguments}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Slice n cases of an eventlog

\item \code{grouped_eventlog}: Slice grouped eventlog: take slice of cases from each group.
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{slice}}}
}}

