% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.eventlog.R
\name{slice_activities}
\alias{slice_activities}
\alias{slice_activities.eventlog}
\alias{slice_activities.grouped_eventlog}
\title{Slice Activities}
\usage{
slice_activities(.data, ...)

\method{slice_activities}{eventlog}(.data, ...)

\method{slice_activities}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Slice index}
}
\description{
Take a slice of activity instances from event log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Take a slice of activity instances from event log

\item \code{grouped_eventlog}: Take a slice of activity instances from grouped event log
}}

