% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.eventlog.R
\name{slice_events}
\alias{slice_events}
\alias{slice_events.eventlog}
\alias{slice_events.grouped_eventlog}
\title{Slice Events}
\usage{
slice_events(.data, ...)

\method{slice_events}{eventlog}(.data, ...)

\method{slice_events}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Slice index}
}
\description{
Take a slice of events from event log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Take a slice of events from event log

\item \code{grouped_eventlog}: Take a slice of events from grouped event log
}}

