% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\docType{import}
\name{slice_sample}
\alias{slice_sample}
\alias{slice_sample.log}
\alias{slice_sample.grouped_log}
\title{Sample function for logs}
\usage{
\method{slice_sample}{log}(.data, ..., n, prop, weight_by = NULL, replace = FALSE)

\method{slice_sample}{grouped_log}(.data, ..., n, prop, weight_by = NULL, replace = FALSE)
}
\section{Methods (by class)}{
\itemize{
\item \code{slice_sample(log)}: Sample \code{n} cases of a \code{\link{log}}.

\item \code{slice_sample(grouped_log)}: Sample \code{n} cases from a \code{\link{grouped_log}}.

}}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:slice]{slice_sample}}}
}}

