% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.eventlog.R
\docType{import}
\name{summarize}
\alias{summarize}
\alias{summarise}
\alias{summarise.eventlog}
\alias{summarise.grouped_eventlog}
\title{Summarize event log}
\usage{
\method{summarise}{eventlog}(.data, ...)

\method{summarise}{grouped_eventlog}(.data, ...)
}
\arguments{
\item{.data}{Eventlog}

\item{...}{Name-value pairs of summary functions}
}
\value{
The summarize function returns a tible, no event log.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Summarize eventlog

\item \code{grouped_eventlog}: Summarize eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{summarise}}}
}}

