% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Timestamp classifier}
\usage{
timestamp(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the  timestamp classifier of an object of class \code{eventlog}
}
\examples{


data(example_log)
timestamp(example_log)


}
\seealso{
\code{\link{eventlog}}
}

