% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamps.R
\name{timestamps}
\alias{timestamps}
\alias{timestamps.eventlog}
\alias{timestamps.eventlog_mapping}
\alias{timestamps.activitylog}
\alias{timestamps.activitylog_mapping}
\title{Timestamp classifiers}
\usage{
timestamps(x)

\method{timestamps}{eventlog}(x)

\method{timestamps}{eventlog_mapping}(x)

\method{timestamps}{activitylog}(x)

\method{timestamps}{activitylog_mapping}(x)
}
\arguments{
\item{x}{Object of class \code{\link{activitylog}}, or \code{\link{mapping}}.}
}
\description{
Get the  timestamps classifier of an object of class \code{activitylog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Retrieve timestamp identifier from eventlog

\item \code{eventlog_mapping}: Retrieve timestamp identifier from eventlog mapping

\item \code{activitylog}: Retrieve timestamp identifier from activitylog

\item \code{activitylog_mapping}: Retrieve timestamp identifier from activitylog mapping
}}

\seealso{
\code{\link{activitylog}}, \code{\link{mapping}}

Other Classifiers: 
\code{\link{activity_id}()},
\code{\link{activity_instance_id}()},
\code{\link{case_id}()},
\code{\link{lifecycle_id}()},
\code{\link{resource_id}()},
\code{\link{set_activity_id}()},
\code{\link{set_activity_instance_id}()},
\code{\link{set_case_id}()},
\code{\link{set_lifecycle_id}()},
\code{\link{set_resource_id}()},
\code{\link{set_timestamp}()},
\code{\link{timestamp}()}
}
\concept{Classifiers}
