% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_list.R
\name{trace_list}
\alias{trace_list}
\alias{trace_list.eventlog}
\alias{trace_list.activitylog}
\alias{trace_list.grouped_log}
\title{Trace list}
\usage{
trace_list(log, ...)

\method{trace_list}{eventlog}(log, ...)

\method{trace_list}{activitylog}(log, ...)

\method{trace_list}{grouped_log}(log, ...)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{...}{Other arguments. Currently not used.}
}
\description{
Construct trace list
}
\section{Methods (by class)}{
\itemize{
\item \code{trace_list(eventlog)}: Construct trace list for event log

\item \code{trace_list(activitylog)}: Construct trace list for activity log

\item \code{trace_list(grouped_log)}: Construct list of traces for grouped log

}}
