% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_list.R
\name{trace_list}
\alias{trace_list}
\alias{trace_list.eventlog}
\alias{trace_list.activitylog}
\alias{trace_list.grouped_log}
\title{Trace list}
\usage{
trace_list(log, ..., eventlog = deprecated())

\method{trace_list}{eventlog}(log, ..., eventlog = deprecated())

\method{trace_list}{activitylog}(log, ..., eventlog = deprecated())

\method{trace_list}{grouped_log}(log, ..., eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Other arguments. Currently not used.}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
Construct trace list
}
\section{Methods (by class)}{
\itemize{
\item \code{trace_list(eventlog)}: Construct trace list for event log

\item \code{trace_list(activitylog)}: Construct trace list for activity log

\item \code{trace_list(grouped_log)}: Construct list of traces for grouped log

}}
