% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{traces}
\alias{traces}
\alias{traces.log}
\alias{traces.grouped_log}
\title{Traces}
\usage{
traces(log, ..., eventlog = deprecated())

\method{traces}{log}(log, ..., eventlog = deprecated())

\method{traces}{grouped_log}(log, ..., eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{eventlog}} or \code{\link{activitylog}}.}

\item{...}{Deprecated arguments}

\item{eventlog}{Deprecated; please use \code{log} instead.}
}
\description{
\code{traces} computes the different activity sequences of an event log
together with their absolute and relative frequencies.
Activity sequences are based on the start timestamp of activities.
}
\section{Methods (by class)}{
\itemize{
\item \code{log}: Construct traces list for eventlog

\item \code{grouped_log}: Construct list of traces for grouped log
}}

\seealso{
\code{\link{cases}}, \code{\link{eventlog}}
}
