% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{traces}
\alias{traces}
\title{Traces}
\usage{
traces(eventlog, output_traces = TRUE, output_cases = FALSE)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{output_traces, output_cases}{Logicals specifying what should be
returned, a list of traces or a list of cases. If both are TRUE, a list of
both is returned.}
}
\description{
\code{traces} computes the different activity sequences of an event log
together with their absolute and relative frequencies.
Activity sequences are based on the start timestamp of activities.
}
\examples{

data(example_log)
traces(example_log)

}
\seealso{
\code{\link{cases}}, \code{\link{eventlog}}
}

