% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup.R
\docType{import}
\name{ungroup}
\alias{ungroup}
\alias{ungroup.activitylog}
\title{Ungroup log}
\usage{
\method{ungroup}{activitylog}(x, ...)
}
\arguments{
\item{x}{Activitylog}

\item{...}{variables to remove from grouping}
}
\section{Methods (by class)}{
\itemize{
\item \code{activitylog}: Ungroup columns in eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:group_by]{ungroup}}}
}}

