% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.eventlog.R, R/ungroup_eventlog.R
\name{ungroup_eventlog}
\alias{ungroup_eventlog}
\alias{ungroup_eventlog.eventlog}
\title{Ungroup event log}
\usage{
ungroup_eventlog(eventlog)

\method{ungroup_eventlog}{eventlog}(eventlog)

ungroup_eventlog(eventlog)

\method{ungroup_eventlog}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{Eventlog}
}
\description{
Remove groups from event log

Remove groups from event log
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Remove groups from event log

\item \code{eventlog}: Remove groups from event log
}}

