% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite.eventlog.R
\docType{import}
\name{unite}
\alias{unite}
\alias{unite.eventlog}
\alias{unite.grouped_eventlog}
\title{Unite multiple columns into one.}
\usage{
\method{unite}{eventlog}(data, col, ..., sep = "_", remove = T)

\method{unite}{grouped_eventlog}(data, col, ..., sep = "_", remove = T)
}
\arguments{
\item{col}{The name of the new column (character value)}

\item{...}{Columns to be united.}

\item{sep}{Seperator}

\item{remove}{Remove original columns}

\item{.data}{Eventlog}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Unite columns in eventlog

\item \code{grouped_eventlog}: Unite columns in grouped eventlog
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyr}{\code{\link[tidyr]{unite}}}
}}

